update ha action
