+++ cmd/action/update/cluster/hacontrolplane/command.go
@@ -1,0 +1,47 @@
+package hacontrolplane
+
+import (
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	"github.com/spf13/cobra"
+)
+
+const (
+	name  = "ha"
+	short = "Upgrade the Tenant Cluster to HA"
+	long  = `Upgrade the Tenant Cluster to HA.
+
+	* Fetch the Cluster CR.
+	* Set replicas to 3.
+	* Update the Cluster CR in the Control Plane.
+
+	`
+)
+
+type Config struct {
+	Logger micrologger.Logger
+}
+
+func New(config Config) (*cobra.Command, error) {
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+
+	f := &flag{}
+
+	r := &runner{
+		flag:   f,
+		logger: config.Logger,
+	}
+
+	c := &cobra.Command{
+		Use:   name,
+		Short: short,
+		Long:  long,
+		RunE:  r.Run,
+	}
+
+	f.Init(c)
+
+	return c, nil
+}
