+++ cmd/cl006/zz_generated.command.go
@@ -9,7 +9,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac000"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac003"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac004"
 )
 
@@ -51,20 +50,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
-	var ac003Cmd *cobra.Command
-	{
-		c := ac003.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac003Cmd, err = ac003.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var ac004Cmd *cobra.Command
 	{
 		c := ac004.Config{
@@ -98,7 +83,6 @@ func New(config Config) (*cobra.Command, error) {
 	f.Init(c)
 
 	c.AddCommand(ac000Cmd)
-	c.AddCommand(ac003Cmd)
 	c.AddCommand(ac004Cmd)
 
 	return c, nil
