--- pkg/action/cl006/ac003/zz_generated.error.go
@@ -1,14 +1,0 @@
-package ac003
-
-import (
-	"github.com/giantswarm/microerror"
-)
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
